add dynamic download of dependencies
- The way spacy models are installed is either through direct gitlink or
through the spacy download module. While trying to publish to testpypi
there was no problem since it allowed gitlink as dependencies, however
pypi repo does not allow direct dependencies such as gitlinks. These
made it impossible to publish a package in pypi with these models as
dependencies for the package. To go around this problem we are now
downloading through the spacy cli download function the modules needed
when we request a module that is not downloaded yet.